ready to submit

+++ table_transactions/create_table_transactions.sql
@@ -2,6 +2,6 @@ CREATE TABLE transactions(
   id INT PRIMARY KEY AUTO_INCREMENT,
   user_id INT NOT NULL,
   datetime_transaction DATETIME NOT NULL,
-  total_price INT NOT NULL DEFAULT (0),
+  total_price FLOAT NOT NULL DEFAULT (0),
   FOREIGN KEY (user_id) REFERENCES users(id)
   );
